+++ extension/js/message-handler.js
@@ -207,6 +207,7 @@ defaultSettings['fixCancer']                    = 'true';
 //defaultSettings['adjustAfterLoad']              = 'true';
 defaultSettings['enableSOAPLink']               = 'true';
 defaultSettings['hidePostButtonInThread']       = 'false';
+defaultSettings['removeOwnReport']              = 'true';
 defaultSettings['collapseTldrQuotes']           = 'false';
 defaultSettings['showLastThreePages']           = 'false';
 defaultSettings['postsPerPage']                 = 'default';