+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog List
 + Allowed users to specify maximum dimensions when restricting image size
 * Fixed right "Top" button not being on any page, not just when page navigator was visible.
 * Fix imgur links to larger sizes of imgur images posted on the forums being broken
++ Add setting to remove report button on own posts
 
 ### v2.0.10 (12/25/2013)